backend/requirements.txt
absl-py==2.0.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
blinker==1.8.2
build==1.2.1
cachetools==5.3.2
certifi==2024.7.4
charset-normalizer==3.3.2
chess==1.10.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.7
distro==1.9.0
easyocr==1.7.1
executing==2.0.1
fastapi==0.112.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.5.4
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
gotrue==2.8.0
groq==0.8.0
grpcio==1.59.2
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
imageio==2.35.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.14.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.110
lazy_loader==0.4
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
neo4j==5.24.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt-einsum==3.3.0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.1.3
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==3.11.0
pluggy==1.5.0
postgrest==0.16.11
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
pytest==8.3.2
python-bidi==0.6.0
python-chess==1.999
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
realtime==2.0.2
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.10
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.30.5
uvloop==0.19.0
virtualenv==20.24.6
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.7
zipp==3.20.0
